requirements.txt
torch
pyyaml
tqdm
pillow
numpy
torchvision
opencv-python
torchaudio
scipy
